Prior compilation after removing templates from matrices
