+++ src/matrices/cc_matrix.cpp
@@ -1,8 +1,7 @@
-// IN THIS FILE: The implementation of a concrete class to store and
-// work with matrices. This implementation makes use of Armadillo's
-// library, thus this is only a wrap for Armadillo's methods
+// IN THIS FILE: Implementation of a concrete class to represent
+// matrices. This is the simplest implementation
 
-#include "cc_matrix_armadillo.tpl.h"
+#include "cc_matrix.h"
 
 namespace scicellxx
 {
@@ -10,9 +9,8 @@ namespace scicellxx
  // ===================================================================
  // Empty constructor
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo()
-  : ACMatrix<T>()
+ CCMatrix::CCMatrix() 
+  : ACMatrix()
  {
   // Delete any data in memory
   clean_up();
@@ -21,61 +19,32 @@ namespace scicellxx
  // ===================================================================
  // Constructor to create an m X n matrix.
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(const unsigned long m, const unsigned long n)
-  : ACMatrix<T>(m, n)
+ CCMatrix::CCMatrix(const unsigned long m, const unsigned long n)
+  : ACMatrix(m, n)
  {
-  // Allocate memory
   allocate_memory(m, n);
-  
-  // DELETE DELETE DELETE DELETE
-  // // Delete any data in memory
-  // clean_up();
-  
-  // // Create an Armadillo's matrix
-  // Arma_matrix_pt = new arma::Mat<T>(m, n);
-  // DELETE DELETE DELETE DELETE
  }
  
  // ===================================================================
  // Constructor where we pass the data for the matrix of size m X n
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(T *matrix_pt,
+ CCMatrix::CCMatrix(Real *matrix_pt,
                        const unsigned long m,
                        const unsigned long n)
- : ACMatrix<T>(m, n)
+ : ACMatrix(m, n)
  {
   // Copy the data from the input vector to the Matrix_pt vector
   set_matrix(matrix_pt, m, n);
  }
  
  // ===================================================================
- // Constructor that creates an Armadillo's matrix from a CCMatrix
+ // Constructor a matrix from a CCVector
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(CCMatrix<T> &matrix)
-  : ACMatrix<T>()
- {
-  // Get the pointer to the matrix data
-  T *matrix_pt = matrix.matrix_pt();
-  // Get the dimension of the new matrix
-  unsigned long m = matrix.n_rows();
-  unsigned long n = matrix.n_columns();
-  
-  // Copy the data from the vector to the Matrix_pt vector
-  set_matrix(matrix_pt, m, n);
- }
- 
- // ===================================================================
- // Constructor that creates an Armadillo's matrix from a CCVector
- // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(CCVector<T> &vector)
-  : ACMatrix<T>()
+ CCMatrix::CCMatrix(CCVector &vector)
+  : ACMatrix()
  {
   // Get the pointer to the vector data
-  T *vector_pt = vector.vector_pt();
+  Real *vector_pt = vector.vector_pt();
   // Compute the dimension of the new matrix by checking whether the
   // vector is transposed or not
   unsigned long m = 0;
@@ -92,55 +61,22 @@ namespace scicellxx
    }
   // Copy the data from the vector to the Matrix_pt vector
   set_matrix(vector_pt, m, n);
-  
- }
-
- // ===================================================================
- // Constructor that creates an Armadillo's matrix from a CCVectorArmadillo
- // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(CCVectorArmadillo<T> &vector)
- {
-  // Compute the dimension of the new matrix by checking whether the
-  // vector is transposed or not
-  unsigned long m = 0;
-  unsigned long n = 0;
-  if (!vector.is_column_vector()) // a row vector
-   {
-    m = 1;
-    n = vector.n_values();
-   }
-  else // a column vector
-   {
-    m = vector.n_values();
-    n = 1;
-   }
-  set_matrix(vector.arma_vector_pt(), m, n);
  }
  
  // ===================================================================
  // Copy constructor
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::CCMatrixArmadillo(const CCMatrixArmadillo<T> &copy)
-  : ACMatrix<T>(copy.n_rows(), copy.n_columns())
+ CCMatrix::CCMatrix(const CCMatrix &copy)
+  : ACMatrix(copy.n_rows(), copy.n_columns())
  {
-  // Clean any possible previously allocated memory
-  clean_up();
-  
-  // Call the copy constructor of Armadillo
-  Arma_matrix_pt = new arma::Mat<T>(*(copy.arma_matrix_pt()));
-  
-  // Mark the matrix as having its own memory
-  this->Is_own_memory_allocated = true;
-  
+  // Copy the data from the copy matrix to the Matrix_pt vector
+  set_matrix(copy.matrix_pt(), this->NRows, this->NColumns);
  }
  
  // ===================================================================
  // Empty destructor
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>::~CCMatrixArmadillo()
+ CCMatrix::~CCMatrix()
  {
   // Deallocate memory
   clean_up();
@@ -149,11 +85,10 @@ namespace scicellxx
  // ===================================================================
  // Assignment operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>& CCMatrixArmadillo<T>::operator=(const CCMatrixArmadillo<T> &source_matrix)
+ CCMatrix& CCMatrix::operator=(const CCMatrix &source_matrix)
  {
   // Clean-up and set values
-  set_matrix(source_matrix.arma_matrix_pt(),
+  set_matrix(source_matrix.matrix_pt(),
              source_matrix.n_rows(),
              source_matrix.n_columns());
   
@@ -164,8 +99,7 @@ namespace scicellxx
  // ===================================================================
  // += operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>& CCMatrixArmadillo<T>::operator+=(const CCMatrixArmadillo<T> &matrix)
+ CCMatrix& CCMatrix::operator+=(const CCMatrix &matrix)
  {
   // Call the method to perform the addition
   add_matrix(matrix, *this);
@@ -176,8 +110,7 @@ namespace scicellxx
  // ===================================================================
  // -= operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T>& CCMatrixArmadillo<T>::operator-=(const CCMatrixArmadillo<T> &matrix)
+ CCMatrix& CCMatrix::operator-=(const CCMatrix &matrix)
  {
   // Call the method to perform the operation
   substract_matrix(matrix, *this);
@@ -188,11 +121,10 @@ namespace scicellxx
  // ===================================================================
  // Add operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T> CCMatrixArmadillo<T>::operator+(const CCMatrixArmadillo<T> &matrix)
+ CCMatrix CCMatrix::operator+(const CCMatrix &matrix)
  {
   // Create a zero matrix where to store the result
-  CCMatrixArmadillo<T> solution(this->NRows, this->NColumns);
+  CCMatrix solution(this->NRows, this->NColumns);
   // Call the method to perform the addition
   add_matrix(matrix, solution);
   // Return the solution matrix
@@ -202,11 +134,10 @@ namespace scicellxx
  // ===================================================================
  // Substraction operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T> CCMatrixArmadillo<T>::operator-(const CCMatrixArmadillo<T> &matrix)
+ CCMatrix CCMatrix::operator-(const CCMatrix &matrix)
  {
   // Create a zero matrix where to store the result
-  CCMatrixArmadillo<T> solution(this->NRows, this->NColumns);
+  CCMatrix solution(this->NRows, this->NColumns);
   // Call the method to perform the operation
   substract_matrix(matrix, solution);
   return solution;
@@ -215,11 +146,10 @@ namespace scicellxx
  // ===================================================================
  // Multiplication operator
  // ===================================================================
- template<class T>
- CCMatrixArmadillo<T> CCMatrixArmadillo<T>::operator*(const CCMatrixArmadillo<T> &right_matrix)
+ CCMatrix CCMatrix::operator*(const CCMatrix &right_matrix)
  { 
   // Create a zero matrix where to store the result
-  CCMatrixArmadillo<T> solution(this->NRows, right_matrix.n_columns());
+  CCMatrix solution(this->NRows, right_matrix.n_columns());
   // Perform the multiplication
   multiply_by_matrix(right_matrix, solution);
   // Return the solution matrix
@@ -227,13 +157,12 @@ namespace scicellxx
  }
  
  // ===================================================================
- // Transforms the input vector to an armadillo matrix class type
- // (virtual such that each derived class has to implement it)
+ // Transforms the input vector to a matrix class type (virtual such
+ // that each derived class has to implement it)
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::set_matrix(const T *matrix_pt,
-                                       const unsigned long m,
-                                       const unsigned long n)
+ void CCMatrix::set_matrix(const Real *matrix_pt,
+                              const unsigned long m,
+                              const unsigned long n)
  {
   // Clean any possible previously allocated memory
   clean_up();
@@ -242,43 +171,21 @@ namespace scicellxx
   this->NRows = m;
   this->NColumns = n;
   
-  // Create an Armadillo's matrix (makes an own copy of the data,
-  // therefore 'matrix_pt' may be deleted safely)
-  Arma_matrix_pt = new arma::Mat<T>(matrix_pt, m, n);
+  // Allocate memory for the matrix
+  Matrix_pt = new Real[m*n];
   
   // Mark the matrix as having its own memory
   this->Is_own_memory_allocated = true;
   
- }
- 
- // ===================================================================
- // Receives an armadillo type matrix
- // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::set_matrix(arma::Mat<T> *arma_matrix_pt,
-                                       const unsigned long m,
-                                       const unsigned long n)
- {
-  // Clean any possible previously allocated memory
-  clean_up();
-  
-  // Set the number of rows and columns
-  this->NRows = m;
-  this->NColumns = n;
-  
-  // Call the copy constructor of Armadillo
-  Arma_matrix_pt = new arma::Mat<T>(*arma_matrix_pt);
-  
-  // Mark the matrix as having its own memory
-  this->Is_own_memory_allocated = true;
+  // Copy the matrix (an element by element copy, uff!!)
+  std::memcpy(Matrix_pt, matrix_pt, m*n*sizeof(Real));
   
  }
  
  // ===================================================================
  // Clean up for any dynamically stored data
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::clean_up()
+ void CCMatrix::clean_up()
  {
   // Check whether the Matrix allocated its own memory
   if (this->Is_own_memory_allocated)
@@ -291,7 +198,7 @@ namespace scicellxx
   else // If empty
    {
     // Set the pointer of the matrix to NULL
-    Arma_matrix_pt = 0;
+    Matrix_pt = 0;
    }
   
  }
@@ -299,16 +206,15 @@ namespace scicellxx
  // ===================================================================
  // Free allocated memory for matrix
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::free_memory_for_matrix()
+ void CCMatrix::free_memory_for_matrix()
  {
   // Is the matrix allowed for deletion. If this method is called from
   // an external source we need to check whether the matrix has been
   // marked for deletion
   if (this->Delete_matrix)
    {
-    delete Arma_matrix_pt;
-    Arma_matrix_pt = 0; 
+    delete [] Matrix_pt;
+    Matrix_pt = 0; 
     
     // Mark the matrix as not having allocated memory
     this->Is_own_memory_allocated=false;
@@ -330,9 +236,8 @@ namespace scicellxx
  // ===================================================================
  // Performs sum of matrices
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::add_matrix(const CCMatrixArmadillo<T> &matrix,
-                                       CCMatrixArmadillo<T> &solution_matrix)
+ void CCMatrix::add_matrix(const CCMatrix &matrix,
+                              CCMatrix &solution_matrix)
  {
   // Check that THIS and the other matrix have memory allocated
   if (!this->Is_own_memory_allocated || !matrix.is_own_memory_allocated())
@@ -367,7 +272,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -384,7 +289,7 @@ namespace scicellxx
      {
       // Error message
       std::ostringstream error_message;
-      error_message << "The dimension of the matrices is not the same:\n"
+      error_message << "The dimensions of the matrices are not the same:\n"
                     << "dim(this) = (" << n_rows << ", "
                     << n_columns << ")\n"
                     << "dim(solution_matrix) = (" << n_rows_solution_matrix
@@ -397,22 +302,27 @@ namespace scicellxx
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
   // Get the matrix pointer of the input matrix
-  arma::Mat<T> *arma_matrix_pt = matrix.arma_matrix_pt();
-  
-  // Perform the operation
-  (*arma_solution_matrix_pt) = (*Arma_matrix_pt) + (*arma_matrix_pt);
+  Real *matrix_pt = matrix.matrix_pt();
+  // Perform the addition
+  for (unsigned long i = 0; i < n_rows; i++)
+   {
+    const unsigned long offset = i*n_columns;
+    for (unsigned long j = 0; j < n_columns; j++)
+     {
+      solution_matrix_pt[offset+j] = Matrix_pt[offset+j] + matrix_pt[offset+j];
+     }
+   }
   
  }
  
  // ===================================================================
  // Performs substraction of matrices
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::substract_matrix(const CCMatrixArmadillo<T> &matrix,
-                                    CCMatrixArmadillo<T> &solution_matrix)
+ void CCMatrix::substract_matrix(const CCMatrix &matrix,
+                                    CCMatrix &solution_matrix)
  {
   // Check that THIS and the other matrix have memory allocated
   if (!this->Is_own_memory_allocated || !matrix.is_own_memory_allocated())
@@ -447,7 +357,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -473,26 +383,31 @@ namespace scicellxx
                              SCICELLXX_CURRENT_FUNCTION,
                              SCICELLXX_EXCEPTION_LOCATION);
      }
-    
+
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
   // Get the matrix pointer of the input matrix
-  arma::Mat<T> *arma_matrix_pt = matrix.arma_matrix_pt();
-  
-  // Perform the operation
-  (*arma_solution_matrix_pt) = (*Arma_matrix_pt) - (*arma_matrix_pt);
-  
+  Real *matrix_pt = matrix.matrix_pt();
+  // Perform the addition
+  for (unsigned long i = 0; i < n_rows; i++)
+   {
+    const unsigned long offset = i*n_columns;
+    for (unsigned long j = 0; j < n_columns; j++)
+     {
+      solution_matrix_pt[offset+j] = Matrix_pt[offset+j] - matrix_pt[offset+j];
+     }
+   }
+ 
  }
 
  // ===================================================================
  // Performs multiplication of matrices
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::multiply_by_matrix(const CCMatrixArmadillo<T> &right_matrix,
-                                               CCMatrixArmadillo<T> &solution_matrix)
+ void CCMatrix::multiply_by_matrix(const CCMatrix &right_matrix,
+                                      CCMatrix &solution_matrix)
  {
   // Check that THIS and the right matrix have memory allocated
   if (!this->Is_own_memory_allocated || !right_matrix.is_own_memory_allocated())
@@ -529,7 +444,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -563,36 +478,64 @@ namespace scicellxx
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
   // Get the matrix pointer of the right matrix
-  arma::Mat<T> *arma_right_matrix_pt = right_matrix.arma_matrix_pt();
-  
-  // Perform the operation
-  (*arma_solution_matrix_pt) = (*Arma_matrix_pt) * (*arma_right_matrix_pt);
+  Real *right_matrix_pt = right_matrix.matrix_pt();
+  // Perform the multiplication
+  for (unsigned long i = 0; i < n_rows_left_matrix; i++)
+   {
+    const unsigned offset_right_matrix = i * n_columns_right_matrix;
+    const unsigned offset_left_matrix = i * n_columns_left_matrix;
+    for (unsigned long j = 0; j < n_columns_right_matrix; j++)
+     {
+      // Initialise
+      solution_matrix_pt[offset_right_matrix+j] = 0;
+      for (unsigned long k = 0; k < n_columns_left_matrix; k++)
+       {
+        solution_matrix_pt[offset_right_matrix+j]+=
+         Matrix_pt[offset_left_matrix+k] * right_matrix_pt[k*n_columns_right_matrix+j];
+       }
+     }
+   }
   
  }
  
  // ===================================================================
  // Computes the transpose and store in the solution matrix
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::transpose(CCMatrixArmadillo<T> &transposed_matrix)
+ void CCMatrix::transpose(CCMatrix &transposed_matrix)
  {
-  // Compute transpose
-  arma::Mat<T> arma_transposed_matrix_pt = Arma_matrix_pt->t();
+  // Get the number of rows and columns of the matrix
+  const unsigned long n_rows = this->NRows;
+  const unsigned long n_columns = this->NColumns;
   
-  transposed_matrix.set_matrix(&arma_transposed_matrix_pt, arma_transposed_matrix_pt.n_rows, arma_transposed_matrix_pt.n_cols);
+  // Create a vector to store the transposed matrix
+  Real *transposed_matrix_pt = new Real[n_rows*n_columns];
+  // Copy the data in the transposed matrix
+  for (unsigned i = 0; i < n_rows; i++)
+   {
+    for (unsigned j = 0; j < n_columns; j++)
+     {    
+      transposed_matrix_pt[j*n_rows+i] = value(i,j);
+     }
+   }
+  
+  // Copy the transposed matrix pointer to the transpose_matrix matrix
+  transposed_matrix.set_matrix(transposed_matrix_pt, n_columns, n_rows);
+  
+  // Delete the temporary transpose matrix pointer
+  delete [] transposed_matrix_pt;
+  
  }
  
  // ===================================================================
  // Computes the transpose and stores it in itself
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::transpose()
+ void CCMatrix::transpose()
  {
   // Check that THIS matrix has memory allocated
-  if (!this->Is_own_memory_allocated)
+  if (!(this->is_own_memory_allocated()))
    {
     // Error message
     std::ostringstream error_message;
@@ -605,20 +548,13 @@ namespace scicellxx
    }
   
   // Transpose itself
-  arma::inplace_trans(*(this->Arma_matrix_pt));
-
-  // Update nrows and columns of transposed_matrix
-  unsigned long NRows_tmp = this->NRows;
-  this->NRows = this->NColumns;
-  this->NColumns = NRows_tmp;
-  
+  this->transpose(*this);
  }
  
  // ===================================================================
  // Get the specified value from the matrix (read-only)
  // ===================================================================
- template<class T>
- const T CCMatrixArmadillo<T>::value(const unsigned long i, const unsigned long j) const
+ const Real CCMatrix::value(const unsigned long i, const unsigned long j) const
  {
 #ifdef SCICELLXX_RANGE_CHECK
   if (!(this->is_own_memory_allocated()))
@@ -657,14 +593,13 @@ namespace scicellxx
    }
 #endif // #ifdef SCICELLXX_RANGE_CHECK
   // Return the value at row i and column j
-  return (*Arma_matrix_pt)(i, j);
+  return Matrix_pt[i*this->NColumns+j];
  }
 
  // ===================================================================
  // Set values in the matrix (write version)
  // ===================================================================
- template<class T>
- T &CCMatrixArmadillo<T>::value(const unsigned long i, const unsigned long j)
+ Real &CCMatrix::value(const unsigned long i, const unsigned long j)
  {
 #ifdef SCICELLXX_RANGE_CHECK
   if (!(this->is_own_memory_allocated()))
@@ -703,14 +638,13 @@ namespace scicellxx
    }
 #endif // #ifdef SCICELLXX_RANGE_CHECK
   // Return the value at row i and column j
-  return (*Arma_matrix_pt)(i, j);
+  return Matrix_pt[i*this->NColumns+j];
  }
  
  // ===================================================================
  /// Permute the rows in the list
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::permute_rows(std::vector<std::pair<unsigned long, unsigned long> > &permute_list)
+ void CCMatrix::permute_rows(std::vector<std::pair<unsigned long, unsigned long> > &permute_list)
  {
   // Get the number of elements in the permute list
   const unsigned n_permute_list = permute_list.size();
@@ -750,7 +684,7 @@ namespace scicellxx
       // Swap values in rows i and j
       
       // Temporarly storage to swap values
-      T tmp = value(j, k);
+      Real tmp = value(j, k);
       // Swap
       value(j, k) = value(i, k);
       value(i, k) = tmp;
@@ -763,8 +697,7 @@ namespace scicellxx
  // ===================================================================
  /// Permute the columns in the list
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::permute_columns(std::vector<std::pair<unsigned long, unsigned long> > &permute_list)
+ void CCMatrix::permute_columns(std::vector<std::pair<unsigned long, unsigned long> > &permute_list)
  {
   // Get the number of elements in the permute list
   const unsigned n_permute_list = permute_list.size();
@@ -804,7 +737,7 @@ namespace scicellxx
       // Swap values in columns i and j
       
       // Temporarly storage to swap values
-      T tmp = value(k, j);
+      Real tmp = value(k, j);
       // Swap
       value(k, j) = value(k, i);
       value(k, i) = tmp;
@@ -817,8 +750,7 @@ namespace scicellxx
  // ===================================================================
  /// Permute rows i and j
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::permute_rows(const unsigned long &i, const unsigned long &j)
+ void CCMatrix::permute_rows(const unsigned long &i, const unsigned long &j)
  {
   // Check that both input columns are within the range
   if (i < this->NRows && j < this->NRows)
@@ -829,7 +761,7 @@ namespace scicellxx
     for (unsigned long k = 0; k < n_columns; k++)
      {
       // Temporarly storage to swap values
-      T tmp = value(j, k);
+      Real tmp = value(j, k);
       // Swap
       value(j, k) = value(i, k);
       value(i, k) = tmp;
@@ -852,8 +784,7 @@ namespace scicellxx
  // ================================================================== 
  // Permute columns i and j
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::permute_columns(const unsigned long &i, const unsigned long &j)
+ void CCMatrix::permute_columns(const unsigned long &i, const unsigned long &j)
  {
   // Check that both input columns are within the range
   if (i < this->NColumns && j < this->NColumns)
@@ -864,7 +795,7 @@ namespace scicellxx
     for (unsigned long k = 0; k < n_rows; k++)
      {
       // Temporarly storage to swap values
-      T tmp = value(k, j);
+      Real tmp = value(k, j);
       // Swap
       value(k, j) = value(k, i);
       value(k, i) = tmp;
@@ -887,8 +818,7 @@ namespace scicellxx
  // ===================================================================
  // Output the matrix
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::output(bool output_indexes) const
+ void CCMatrix::output(bool output_indexes) const
  {
   if (!this->Is_own_memory_allocated)
    {
@@ -909,25 +839,32 @@ namespace scicellxx
         for (unsigned long j = 0; j < this->NColumns; j++)
          {
           std::cout << "(" << i << ", " << j << "): "
-                    << value(i,j) << std::endl; 
+                    << value(i,j)
+                    << std::endl; 
          } // for (j < this->NColumns)
        } // for (i < this->NRows)
      } // if (output_indexes)
-   else
+    else
      {
-      Arma_matrix_pt->print();
+      for (unsigned long i = 0; i < this->NRows; i++)
+       {
+        for (unsigned long j = 0; j < this->NColumns; j++)
+         {
+          std::cout << value(i,j) << " ";
+         } // for (j < this->NColumns)
+        std::cout << std::endl;
+       } // for (i < this->NRows)
      } // else if (output_indexes)
-   
+    
    }
- 
+  
  }
 
  // ===================================================================
  // Output the matrix
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::output(std::ofstream &outfile,
-                                   bool output_indexes) const
+ void CCMatrix::output(std::ofstream &outfile,
+                          bool output_indexes) const
  {
   if (!this->Is_own_memory_allocated)
    {
@@ -948,24 +885,32 @@ namespace scicellxx
         for (unsigned long j = 0; j < this->NColumns; j++)
          {
           outfile << "(" << i << ", " << j << "): "
-                  << value(i,j) << std::endl; 
+                  << value(i,j)
+                  << std::endl; 
          } // for (j < this->NColumns)
        } // for (i < this->NRows)
      } // if (output_indexes)
     else
      {
-      Arma_matrix_pt->print(outfile);
+      for (unsigned long i = 0; i < this->NRows; i++)
+       {
+        for (unsigned long j = 0; j < this->NColumns; j++)
+         {
+          outfile << value(i,j) << " ";
+         } // for (j < this->NColumns)
+        outfile << std::endl;
+       } // for (i < this->NRows)
      } // else if (output_indexes)
-  }
-      
+
+   }
+  
  }
  
  // ===================================================================
  // Allows to create a matrix with the given size but with no data
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::allocate_memory(const unsigned long m,
-                                            const unsigned long n)
+ void CCMatrix::allocate_memory(const unsigned long m,
+                                   const unsigned long n)
  {
   // Clean any possibly stored data
   clean_up();
@@ -978,7 +923,7 @@ namespace scicellxx
   //allocate_memory();
   
   // Allocate memory for the matrix
-  Arma_matrix_pt = new arma::Mat<T>(this->NRows, this->NColumns);
+  Matrix_pt = new Real[this->NRows * this->NColumns];
   
   // Mark the matrix as allocated its own memory
   this->Is_own_memory_allocated=true;
@@ -988,30 +933,28 @@ namespace scicellxx
  // // ===================================================================
  // // Allocates memory to store entries of the matrix
  // // ===================================================================
- // template<class T>
- // void CCMatrixArmadillo<T>::allocate_memory()
+ // void CCMatrix::allocate_memory()
  // {
  //  // Delete any data in memory
  //  clean_up();
   
  //  // Allocate memory for the matrix
- //  Arma_matrix_pt = new arma::Mat<T>(this->NRows, this->NColumns);
+ //  Matrix_pt = new Real[this->NRows * this->NColumns];
   
  //  // Mark the matrix as allocated its own memory
  //  this->Is_own_memory_allocated=true;
  // }
- 
+
  // ===================================================================
  // Fills the matrix with zeroes
  // ===================================================================
- template<class T>
- void CCMatrixArmadillo<T>::fill_with_zeroes()
+ void CCMatrix::fill_with_zeroes()
  {
   // Check that the matrix has memory allocated
   if (this->Is_own_memory_allocated)
    {
     // Fill the matrix with zeroes
-    Arma_matrix_pt->zeros();
+    std::memset(Matrix_pt, 0, this->NRows * this->NColumns * sizeof(Real));
    }
   else
    {
@@ -1036,10 +979,9 @@ namespace scicellxx
  // ===================================================================
  // Performs sum of matrices
  // ===================================================================
- template<class T>
- void add_matrices(const CCMatrixArmadillo<T> &matrix_one,
-                   const CCMatrixArmadillo<T> &matrix_two,
-                   CCMatrixArmadillo<T> &solution_matrix)
+ void add_matrices(const CCMatrix &matrix_one,
+                   const CCMatrix &matrix_two,
+                   CCMatrix &solution_matrix)
  {
   // Check that both matrices have memory allocated
   if (!matrix_one.is_own_memory_allocated() || !matrix_two.is_own_memory_allocated())
@@ -1076,7 +1018,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -1107,24 +1049,30 @@ namespace scicellxx
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
   // Get the matrix pointer of the input matrices
-  arma::Mat<T> *arma_matrix_one_pt = matrix_one.arma_matrix_pt();
-  arma::Mat<T> *arma_matrix_two_pt = matrix_two.arma_matrix_pt();
-  
+  Real *matrix_one_pt = matrix_one.matrix_pt();
+  Real *matrix_two_pt = matrix_two.matrix_pt();
   // Perform the addition
-  (*arma_solution_matrix_pt) = (*arma_matrix_one_pt) + (*arma_matrix_two_pt);
+  for (unsigned long i = 0; i < n_rows_matrix_one; i++)
+   {
+    const unsigned long offset = i*n_columns_matrix_one;
+    for (unsigned long j = 0; j < n_columns_matrix_one; j++)
+     {
+      solution_matrix_pt[offset+j] =
+       matrix_one_pt[offset+j] + matrix_two_pt[offset+j];
+     }
+   }
   
  }
  
  // ===================================================================
  // Performs substraction of matrices
  // ===================================================================
- template<class T>
- void substract_matrices(const CCMatrixArmadillo<T> &matrix_one,
-                         const CCMatrixArmadillo<T> &matrix_two,
-                         CCMatrixArmadillo<T> &solution_matrix)
+ void substract_matrices(const CCMatrix &matrix_one,
+                         const CCMatrix &matrix_two,
+                         CCMatrix &solution_matrix)
  {
   // Check that both matrices have memory allocated
   if (!matrix_one.is_own_memory_allocated() || !matrix_two.is_own_memory_allocated())
@@ -1161,7 +1109,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-  
+
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -1188,28 +1136,34 @@ namespace scicellxx
                              SCICELLXX_CURRENT_FUNCTION,
                              SCICELLXX_EXCEPTION_LOCATION);
      }
-
+    
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
   // Get the matrix pointer of the input matrices
-  arma::Mat<T> *arma_matrix_one_pt = matrix_one.arma_matrix_pt();
-  arma::Mat<T> *arma_matrix_two_pt = matrix_two.arma_matrix_pt();
-  
-  // Perform the addition
-  (*arma_solution_matrix_pt) = (*arma_matrix_one_pt) - (*arma_matrix_two_pt);
+  Real *matrix_one_pt = matrix_one.matrix_pt();
+  Real *matrix_two_pt = matrix_two.matrix_pt();
+  // Perform the substraction
+  for (unsigned long i = 0; i < n_rows_matrix_one; i++)
+   {
+    const unsigned long offset = i*n_columns_matrix_one;
+    for (unsigned long j = 0; j < n_columns_matrix_one; j++)
+     {
+      solution_matrix_pt[offset+j] =
+       matrix_one_pt[offset+j] - matrix_two_pt[offset+j];
+     }
+   }
   
  }
  
  // ===================================================================
  // Performs multiplication of matrices
  // ===================================================================
- template<class T>
- void multiply_matrices(const CCMatrixArmadillo<T> &left_matrix,
-                        const CCMatrixArmadillo<T> &right_matrix,
-                        CCMatrixArmadillo<T> &solution_matrix)
+ void multiply_matrices(const CCMatrix &left_matrix,
+                        const CCMatrix &right_matrix,
+                        CCMatrix &solution_matrix)
  {
   // Check that both matrices have memory allocated
   if (!left_matrix.is_own_memory_allocated() || !right_matrix.is_own_memory_allocated())
@@ -1254,213 +1208,83 @@ namespace scicellxx
     // Allocate memory for the matrix
     solution_matrix.allocate_memory(n_rows_left_matrix, n_columns_right_matrix);
    }
-  else
+  
+  // Check whether the dimension of the solution matrix are correct
+  const unsigned long n_rows_solution_matrix = solution_matrix.n_rows();
+  const unsigned long n_columns_solution_matrix = solution_matrix.n_columns();
+  if (n_rows_left_matrix != n_rows_solution_matrix ||
+      n_columns_right_matrix != n_columns_solution_matrix)
    {
-    // Check whether the dimension of the solution matrix are correct
-    const unsigned long n_rows_solution_matrix = solution_matrix.n_rows();
-    const unsigned long n_columns_solution_matrix = solution_matrix.n_columns();
-    if (n_rows_left_matrix != n_rows_solution_matrix ||
-        n_columns_right_matrix != n_columns_solution_matrix)
-     {
-      // Error message
-      std::ostringstream error_message;
-      error_message << "The dimension of the solution matrix is not appropiate for\n"
-                    << "the operation:\n"
-                    << "dim(left_matrix) = (" << n_rows_left_matrix << ", "
-                    << n_columns_left_matrix << ")\n"
-                    << "dim(right_matrix) = (" << n_rows_right_matrix << ", "
-                    << n_columns_right_matrix << ")\n"
-                    << "dim(solution_matrix) = (" << n_rows_solution_matrix
-                    << ", " << n_columns_solution_matrix << ")\n" << std::endl;
-      throw SciCellxxLibError(error_message.str(),
-                             SCICELLXX_CURRENT_FUNCTION,
-                             SCICELLXX_EXCEPTION_LOCATION);
-     }
-    
+    // Error message
+    std::ostringstream error_message;
+    error_message << "The dimension of the solution matrix is not appropiate for\n"
+                  << "the operation:\n"
+                  << "dim(left_matrix) = (" << n_rows_left_matrix << ", "
+                  << n_columns_left_matrix << ")\n"
+                  << "dim(right_matrix) = (" << n_rows_right_matrix << ", "
+                  << n_columns_right_matrix << ")\n"
+                  << "dim(solution_matrix) = (" << n_rows_solution_matrix
+                  << ", " << n_columns_solution_matrix << ")\n" << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
    }
   
   // Get the matrix pointer of the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
-  // Get the matrix pointer of the input matrices
-  arma::Mat<T> *arma_matrix_one_pt = left_matrix.arma_matrix_pt();
-  arma::Mat<T> *arma_matrix_two_pt = right_matrix.arma_matrix_pt();
-  
-  // Perform the addition
-  (*arma_solution_matrix_pt) = (*arma_matrix_one_pt) * (*arma_matrix_two_pt);
-  
+  // Get the matrix pointer of the left matrix
+  Real *left_matrix_pt = left_matrix.matrix_pt();
+  // Get the matrix pointer of the right matrix
+  Real *right_matrix_pt = right_matrix.matrix_pt();
+  // Perform the multiplication
+  for (unsigned long i = 0; i < n_rows_left_matrix; i++)
+   {
+    const unsigned offset_left_matrix = i * n_columns_left_matrix;
+    const unsigned offset_right_matrix = i * n_columns_right_matrix;
+    for (unsigned long j = 0; j < n_columns_right_matrix; j++)
+     {
+      // Initialise
+      solution_matrix_pt[offset_right_matrix+j] = 0;
+      for (unsigned long k = 0; k < n_columns_left_matrix; k++)
+       {
+        solution_matrix_pt[offset_right_matrix+j]+=
+         left_matrix_pt[offset_left_matrix+k] * right_matrix_pt[k*n_columns_right_matrix+j];
+       }
+     }
+   }
  }
  
  // ================================================================
  // Concatenate matrices horizontally
  // ================================================================
- template<class T>
-  void concatenate_matrices_horizontally(const CCMatrixArmadillo<T> &left_matrix,
-                                         const CCMatrixArmadillo<T> &right_matrix,
-                                         CCMatrixArmadillo<T> &concatenated_matrix)
+ void concatenate_matrices_horizontally(const CCMatrix &left_matrix,
+                                        const CCMatrix &right_matrix,
+                                        CCMatrix &concatenated_matrix)
  {
-  // Check that both matrices have memory allocated
-  if (!left_matrix.is_own_memory_allocated() || !right_matrix.is_own_memory_allocated())
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "One of the matrices to operate with has no memory allocated\n"
-                  << "left_matrix.is_own_memory_allocated() = "
-                  << left_matrix.is_own_memory_allocated() << "\n"
-                  << "right_matrix.is_own_memory_allocated() = "
-                  << right_matrix.is_own_memory_allocated() << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Check whether the number of rows of the matrices are the same
-  const unsigned long n_rows_left_matrix = left_matrix.n_rows();
-  const unsigned long n_rows_right_matrix = right_matrix.n_rows();
-  if (n_rows_left_matrix != n_rows_right_matrix)
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "The number of rows of the matrices is not the same:\n"
-                  << "n_rows(left_matrix) = (" << n_rows_left_matrix << ")\n"
-                  << "n_rows(right_matrix) = (" << n_rows_right_matrix << ")\n"
-                  << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Get the number of columns of each matrix and compute the new
-  // number of columns for the new matrix
-  const unsigned long n_columns_left_matrix = left_matrix.n_columns();
-  const unsigned long n_columns_right_matrix = right_matrix.n_columns();
-  const unsigned long n_columns_new_concatenated_matrix =
-   n_columns_left_matrix + n_columns_right_matrix;
-  const unsigned long n_rows_new_concatenated_matrix = n_rows_left_matrix;
-  
-  // Check whether the concatenated matrix has allocated memory,
-  // otherwise allocate it here!!!
-  if (!concatenated_matrix.is_own_memory_allocated())
-   {
-    // Allocate memory for the matrix
-    concatenated_matrix.allocate_memory(n_rows_new_concatenated_matrix,
-                                        n_columns_new_concatenated_matrix);
-   }
-  
-  // Check whether the dimension of the concatenated matrix is correct
-  const unsigned long n_rows_concatenated_matrix = concatenated_matrix.n_rows();
-  const unsigned long n_columns_concatenated_matrix = concatenated_matrix.n_columns();
-  if (n_rows_concatenated_matrix != n_rows_new_concatenated_matrix ||
-      n_columns_concatenated_matrix != n_columns_new_concatenated_matrix)
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "The dimension of the matrix is not as expected:\n"
-                  << "dim(concatenated_matrix) = (" << n_rows_concatenated_matrix << ", "
-                  << n_columns_concatenated_matrix << ")\n"
-                  << "dim(expected_matrix) = (" << n_rows_new_concatenated_matrix
-                  << ", " << n_columns_new_concatenated_matrix << ")\n" << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Get the matrix pointer of the concatenated matrix
-  arma::Mat<T> *arma_concatenated_matrix_pt = concatenated_matrix.arma_matrix_pt();
-  
-  // Get the matrix pointer of the input matrices
-  arma::Mat<T> *arma_left_matrix_pt = left_matrix.arma_matrix_pt();
-  arma::Mat<T> *arma_right_matrix_pt = right_matrix.arma_matrix_pt();
-  
-  // Perform the concatenation of rows/horizontal_concatenation
-  (*arma_concatenated_matrix_pt) = arma::join_rows((*arma_left_matrix_pt), (*arma_right_matrix_pt));
-  
+  // Error message
+  std::ostringstream error_message;
+  error_message << "This method is not currently implemented for this type of matrices"
+                << "Please implemented it yourself" << std::endl;
+  throw SciCellxxLibError(error_message.str(),
+                         SCICELLXX_CURRENT_FUNCTION,
+                         SCICELLXX_EXCEPTION_LOCATION);
  }
  
  // ================================================================
  // Concatenate matrices vertically
  // ================================================================
- template<class T>
-  void concatenate_matrices_vertically(const CCMatrixArmadillo<T> &upper_matrix,
-                                       const CCMatrixArmadillo<T> &lower_matrix,
-                                       CCMatrixArmadillo<T> &concatenated_matrix)
+ void concatenate_matrices_vertically(const CCMatrix &upper_matrix,
+                                      const CCMatrix &lower_matrix,
+                                      CCMatrix &concatenated_matrix)
  {
-  // Check that both matrices have memory allocated
-  if (!upper_matrix.is_own_memory_allocated() || !lower_matrix.is_own_memory_allocated())
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "One of the matrices to operate with has no memory allocated\n"
-                  << "upper_matrix.is_own_memory_allocated() = "
-                  << upper_matrix.is_own_memory_allocated() << "\n"
-                  << "lower_matrix.is_own_memory_allocated() = "
-                  << lower_matrix.is_own_memory_allocated() << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Check whether the number of columns of the matrices are the same
-  const unsigned long n_columns_upper_matrix = upper_matrix.n_columns();
-  const unsigned long n_columns_lower_matrix = lower_matrix.n_columns();
-  if (n_columns_upper_matrix != n_columns_lower_matrix)
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "The number of columns of the matrices is not the same:\n"
-                  << "n_columns(upper_matrix) = (" << n_columns_upper_matrix << ")\n"
-                  << "n_columns(lower_matrix) = (" << n_columns_lower_matrix << ")\n"
-                  << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Get the number of rows of each matrix and compute the new number
-  // of rows for the new matrix
-  const unsigned long n_rows_upper_matrix = upper_matrix.n_rows();
-  const unsigned long n_rows_lower_matrix = lower_matrix.n_rows();
-  const unsigned long n_rows_new_concatenated_matrix =
-   n_rows_upper_matrix + n_rows_lower_matrix;
-  const unsigned long n_columns_new_concatenated_matrix = n_columns_upper_matrix;
-  
-  // Check whether the concatenated matrix has allocated memory,
-  // otherwise allocate it here!!!
-  if (!concatenated_matrix.is_own_memory_allocated())
-   {
-    // Allocate memory for the matrix
-    concatenated_matrix.allocate_memory(n_rows_new_concatenated_matrix,
-                                        n_columns_new_concatenated_matrix);
-   }
-  
-  // Check whether the dimension of the concatenated matrix is correct
-  const unsigned long n_rows_concatenated_matrix = concatenated_matrix.n_rows();
-  const unsigned long n_columns_concatenated_matrix = concatenated_matrix.n_columns();
-  if (n_rows_concatenated_matrix != n_rows_new_concatenated_matrix ||
-      n_columns_concatenated_matrix != n_columns_new_concatenated_matrix)
-   {
-    // Error message
-    std::ostringstream error_message;
-    error_message << "The dimension of the new matrix is not as expected:\n"
-                  << "dim(concatenated_matrix) = (" << n_rows_concatenated_matrix << ", "
-                  << n_columns_concatenated_matrix << ")\n"
-                  << "dim(expected_matrix) = (" << n_rows_new_concatenated_matrix
-                  << ", " << n_columns_new_concatenated_matrix << ")\n" << std::endl;
-    throw SciCellxxLibError(error_message.str(),
-                           SCICELLXX_CURRENT_FUNCTION,
-                           SCICELLXX_EXCEPTION_LOCATION);
-   }
-  
-  // Get the matrix pointer of the concatenated matrix
-  arma::Mat<T> *arma_concatenated_matrix_pt = concatenated_matrix.arma_matrix_pt();
-  
-  // Get the matrix pointer of the input matrices
-  arma::Mat<T> *arma_upper_matrix_pt = upper_matrix.arma_matrix_pt();
-  arma::Mat<T> *arma_lower_matrix_pt = lower_matrix.arma_matrix_pt();
-  
-  // Perform the concatenation of rows/horizontal_concatenation
-  (*arma_concatenated_matrix_pt) = arma::join_cols((*arma_upper_matrix_pt), (*arma_lower_matrix_pt));
-  
+  // Error message
+  std::ostringstream error_message;
+  error_message << "This method is not currently implemented for this type of matrices"
+                << "Please implemented it yourself" << std::endl;
+  throw SciCellxxLibError(error_message.str(),
+                         SCICELLXX_CURRENT_FUNCTION,
+                         SCICELLXX_EXCEPTION_LOCATION);
  }
  
  // ================================================================
@@ -1469,13 +1293,11 @@ namespace scicellxx
 
  // ================================================================
  // Multiply vector times vector (if you want to perform dot product
- // use the dot() method defined in the cc_vector_armadillo.tpl.h file
- // instead)
+ // use the dot() method defined in the cc_vector.tpl.h file instead)
  // ================================================================
- template<class T>
- void multiply_vector_times_vector(const CCVectorArmadillo<T> &left_vector,
-                                   const CCVectorArmadillo<T> &right_vector,
-                                   CCMatrixArmadillo<T> &solution_matrix)
+ void multiply_vector_times_vector(const CCVector &left_vector,
+                                   const CCVector &right_vector,
+                                   CCMatrix &solution_matrix)
  {
   // Check that the left and the right vectors have memory allocated
   if (!left_vector.is_own_memory_allocated() || !right_vector.is_own_memory_allocated())
@@ -1534,7 +1356,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the solution matrix has allocated memory, otherwise
   // allocate it here!!!
   if (!solution_matrix.is_own_memory_allocated())
@@ -1564,28 +1386,41 @@ namespace scicellxx
                              SCICELLXX_CURRENT_FUNCTION,
                              SCICELLXX_EXCEPTION_LOCATION);
      }
-    
+
    }
   
-  // Get the vector pointers of the input vectors
-  arma::Mat<T> *arma_vector_left_pt = left_vector.arma_vector_pt();
-  arma::Mat<T> *arma_vector_right_pt = right_vector.arma_vector_pt();
+  // Get the matrix pointer of the solution matrix
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
   
-  // Get the matrix pointer to the solution matrix
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  // Get both vectors and multiply them
+  Real *left_vector_pt = left_vector.vector_pt();
+  Real *right_vector_pt = right_vector.vector_pt();
   
   // Perform the multiplication
-  (*arma_solution_matrix_pt) = (*arma_vector_left_pt) * (*arma_vector_right_pt);
+  for (unsigned long i = 0; i < n_rows_left_vector; i++)
+   {
+    const unsigned offset_left_vector = i * n_columns_left_vector;
+    const unsigned offset_right_vector = i * n_columns_right_vector;
+    for (unsigned long j = 0; j < n_columns_right_vector; j++)
+     {
+      // Initialise
+      solution_matrix_pt[offset_right_vector+j] = 0;
+      for (unsigned long k = 0; k < n_columns_left_vector; k++)
+       {
+        solution_matrix_pt[offset_right_vector+j]+=
+         left_vector_pt[offset_left_vector+k] * right_vector_pt[k*n_columns_right_vector+j];
+       }
+     }
+   }
   
  }
  
  // ================================================================
  // Multiply vector times matrix
  // ================================================================
- template<class T>
- void multiply_vector_times_matrix(const CCVectorArmadillo<T> &vector,
-                                   const CCMatrixArmadillo<T> &matrix,
-                                   CCMatrixArmadillo<T> &solution_matrix)
+ void multiply_vector_times_matrix(const CCVector &vector,
+                                   const CCMatrix &matrix,
+                                   CCMatrix &solution_matrix)
  {
   // Check that the vector and the matrix have memory allocated
   if (!vector.is_own_memory_allocated() || !matrix.is_own_memory_allocated())
@@ -1666,26 +1501,41 @@ namespace scicellxx
                              SCICELLXX_CURRENT_FUNCTION,
                              SCICELLXX_EXCEPTION_LOCATION);
      }
-    
+
    }
   
-  // Get the vector and the matrix pointer
-  arma::Mat<T> *arma_vector_pt = vector.arma_vector_pt();
-  arma::Mat<T> *arma_matrix_pt = matrix.arma_matrix_pt();
-  arma::Mat<T> *arma_solution_matrix_pt = solution_matrix.arma_matrix_pt();
+  // Get the matrix pointer of the solution matrix
+  Real *solution_matrix_pt = solution_matrix.matrix_pt();
+  
+  // Get both the vector and the matrix pointer
+  Real *vector_pt = vector.vector_pt();
+  Real *matrix_pt = matrix.matrix_pt();
   
   // Perform the multiplication
-  (*arma_solution_matrix_pt) = (*arma_vector_pt) * (*arma_matrix_pt);
+  for (unsigned long i = 0; i < n_rows_vector; i++)
+   {
+    const unsigned offset_vector = i * n_columns_vector;
+    const unsigned offset_matrix = i * n_columns_matrix;
+    for (unsigned long j = 0; j < n_columns_matrix; j++)
+     {
+      // Initialise
+      solution_matrix_pt[offset_matrix+j] = 0;
+      for (unsigned long k = 0; k < n_columns_vector; k++)
+       {
+        solution_matrix_pt[offset_matrix+j]+=
+         vector_pt[offset_vector+k] * matrix_pt[k*n_columns_matrix+j];
+       }
+     }
+   }
   
  }
  
  // ================================================================
  // Multiply matrix times vector
  // ================================================================
- template<class T>
- void multiply_matrix_times_vector(const CCMatrixArmadillo<T> &matrix,
-                                   const CCVectorArmadillo<T> &vector,
-                                   CCVectorArmadillo<T> &solution_vector)
+ void multiply_matrix_times_vector(const CCMatrix &matrix,
+                                   const CCVector &vector,
+                                   CCVector &solution_vector)
  {
   // Check that the matrix and the vector have memory allocated
   if (!matrix.is_own_memory_allocated() || !vector.is_own_memory_allocated())
@@ -1701,7 +1551,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check whether the vector is a column vector
   if (!vector.is_column_vector())
    {
@@ -1755,7 +1605,7 @@ namespace scicellxx
    }
   
   // Compute the dimensions for the solution vector
-  const unsigned n_rows_solution_vector = n_rows_matrix;
+  const unsigned n_rows_solution_vector = n_columns_matrix;
   const unsigned n_columns_solution_vector = n_columns_vector;
   
   // Check whether the solution vector has allocated memory, otherwise
@@ -1787,19 +1637,16 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  arma::Mat<T> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
+  Real *solution_vector_pt = solution_vector.vector_pt();
   
   // Get both the vector and the matrix pointer
-  arma::Mat<T> *arma_vector_pt = vector.arma_vector_pt();
-  arma::Mat<T> *arma_matrix_pt = matrix.arma_matrix_pt();
+  Real *matrix_pt = matrix.matrix_pt();
+  Real *vector_pt = vector.vector_pt();
   
-  // Perform the multiplication
-  (*arma_solution_vector_pt) = (*arma_matrix_pt) * (*arma_vector_pt);  
-  
-#if 0
   // Perform the multiplication
   for (unsigned long i = 0; i < n_rows_solution_vector; i++)
    {
+    const unsigned offset_matrix = i * n_columns_matrix;
     const unsigned offset_vector = i * n_columns_vector;
     // We can skip one loop since there is only one column
     
@@ -1808,10 +1655,10 @@ namespace scicellxx
     for (unsigned long k = 0; k < n_columns_matrix; k++)
      {
       solution_vector_pt[offset_vector]+=
-       matrix(i, k) * vector_pt[k];
+       matrix_pt[offset_matrix+k] * vector_pt[k];
      }
    }
-#endif // #if 0
+  
   
  }
