+++ src/matrices/cc_vector_armadillo.cpp
@@ -1,7 +1,8 @@
-// IN THIS FILE: Implementation of a concrete class to represent
-// vectors. This is the simplest implementation
+// IN THIS FILE: The implementation of a concrete class to store and
+// work with vector. This implementation makes use of Armadillo's
+// library, thus this is only a wrap for Armadillo's methods
 
-#include "cc_vector.tpl.h"
+#include "cc_vector_armadillo.h"
 
 namespace scicellxx
 {
@@ -9,9 +10,8 @@ namespace scicellxx
  // ===================================================================
  // Empty constructor
  // ===================================================================
- template<class T>
- CCVector<T>::CCVector() 
-  : ACVector<T>()
+ CCVectorArmadillo::CCVectorArmadillo()
+  : ACVector()
  {
   // Delete any data in memory
   clean_up();
@@ -20,44 +20,66 @@ namespace scicellxx
  // ===================================================================
  // Constructor to create an n size zero vector (we assume vectors
  // are created as column vectors, if you need a row vector then
- // pass "false" as the second parameter)
+ // pass "true" as the second parameter).
  // ===================================================================
- template<class T>
- CCVector<T>::CCVector(const unsigned long n, bool is_column_vector)
-  : ACVector<T>(n, is_column_vector)
- {
+ CCVectorArmadillo::CCVectorArmadillo(const unsigned long n, bool is_column_vector)
+  : ACVector(n, is_column_vector)
+ {  
   // Allocate memory
-  allocate_memory(n);
+  allocate_memory(n);  
  }
  
  // ===================================================================
  // Constructor where we pass the data for the vector of size n
  // ===================================================================
- template<class T>
- CCVector<T>::CCVector(T *vector_pt, const unsigned long n,
-                       bool is_column_vector)
-  : ACVector<T>(n, is_column_vector)
+ CCVectorArmadillo::CCVectorArmadillo(T *vector_pt, const unsigned long n,
+                                         bool is_column_vector)
+  : ACVector(n, is_column_vector)
  {
   // Copy the data from the input vector to the Vector_pt vector
   set_vector(vector_pt, n, is_column_vector);
  }
  
  // ===================================================================
+ // Constructor that creates an Armadillo's vector from a CCVector
+ // ===================================================================
+ CCVectorArmadillo::CCVectorArmadillo(CCVector &vector)
+ {
+  // Get the pointer to the vector data
+  T *vector_pt = vector.vector_pt();
+  // Get the dimension of the new vector
+  unsigned long n = vector.n_values();
+  
+  // Copy the data from the vector to the Matrix_pt vector
+  set_vector(vector_pt, n, vector.is_column_vector());
+ }
+ 
+ // ===================================================================
  // Copy constructor
  // ===================================================================
- template<class T>
- CCVector<T>::CCVector(const CCVector<T> &copy)
-  : ACVector<T>(copy.n_values(), copy.is_column_vector())
+ CCVectorArmadillo::CCVectorArmadillo(const CCVectorArmadillo &copy)
+  : ACVector(copy.n_values(), copy.is_column_vector())
  {
-  // Copy the data from the input vector to the Vector_pt vector
-  set_vector(copy.vector_pt(), this->NValues, copy.is_column_vector());
+  // Clean any possible previously allocated memory
+  clean_up();
+  
+  // Set the number of values
+  this->NValues = copy.n_values();
+  
+  // Call the copy constructor of Armadillo
+  Arma_vector_pt = new arma::Mat<Real>(*(copy.arma_vector_pt())); 
+  
+  // Mark the matrix as having its own memory
+  this->Is_own_memory_allocated = true;
+  
+  // Set the transposed status
+  this->set_as_column_vector(copy.is_column_vector()); 
  }
  
  // ===================================================================
  // Empty destructor
  // ===================================================================
- template<class T>
- CCVector<T>::~CCVector()
+ CCVectorArmadillo::~CCVectorArmadillo()
  {
   // Deallocate memory
   clean_up();
@@ -65,12 +87,11 @@ namespace scicellxx
  
  // ===================================================================
  // Assignment operator
- // ===================================================================
- template<class T>
- CCVector<T>& CCVector<T>::operator=(const CCVector<T> &source_vector)
+ // =================================================================== 
+ CCVectorArmadillo& CCVectorArmadillo::operator=(const CCVectorArmadillo &source_vector)
  {
   // Clean-up and set values
-  set_vector(source_vector.vector_pt(),
+  set_vector(source_vector.arma_vector_pt(),
              source_vector.n_values(),
              source_vector.is_column_vector());
   // Return this (de-referenced pointer)
@@ -81,8 +102,7 @@ namespace scicellxx
  // ===================================================================
  // += operator
  // ===================================================================
- template<class T>
- CCVector<T>& CCVector<T>::operator+=(const CCVector<T> &vector)
+ CCVectorArmadillo& CCVectorArmadillo::operator+=(const CCVectorArmadillo &vector)
  {  
   // Call the method to perform the addition
   add_vector(vector, *this);
@@ -93,8 +113,7 @@ namespace scicellxx
  // ===================================================================
  // -= operator
  // ===================================================================
- template<class T>
- CCVector<T>& CCVector<T>::operator-=(const CCVector<T> &vector)
+ CCVectorArmadillo& CCVectorArmadillo::operator-=(const CCVectorArmadillo &vector)
  {
   // Call the method to perform the operation
   substract_vector(vector, *this);
@@ -105,11 +124,10 @@ namespace scicellxx
  // ===================================================================
  // Add operator
  // ===================================================================
- template<class T>
- CCVector<T> CCVector<T>::operator+(const CCVector<T> &vector)
+ CCVectorArmadillo CCVectorArmadillo::operator+(const CCVectorArmadillo &vector)
  {
   // Create a zero vector where to store the result
-  CCVector<T> solution(this->NValues);
+  CCVectorArmadillo solution(this->NValues);
   // Call the method to perform the addition
   add_vector(vector, solution);
   // Return the solution vector
@@ -119,27 +137,28 @@ namespace scicellxx
  // ===================================================================
  // Substraction operator
  // ===================================================================
- template<class T>
- CCVector<T> CCVector<T>::operator-(const CCVector<T> &vector)
+ CCVectorArmadillo CCVectorArmadillo::operator-(const CCVectorArmadillo &vector)
  {
   // Create a zero vector where to store the result
-  CCVector<T> solution(this->NValues);
+  CCVectorArmadillo solution(this->NValues);
   // Call the method to perform the operation
   substract_vector(vector, solution);
   return solution;
  }
-  
+ 
+ // HERE HERE HERE Working on this
+ 
  // ===================================================================
  // Multiplication operator (it returns a matrix with the
  // corresponding size, if you require a dot product operation use the
  // dot() method instead
  // ===================================================================
- template<class T>
- CCMatrix<T> CCVector<T>::operator*(const CCVector<T> &vector)
- {  
+ CCMatrixArmadillo CCVectorArmadillo::operator*(const CCVectorArmadillo &vector)
+ {
   // Create two matrices, one from each vector
-  CCMatrix<T> left_matrix(*this);
-  CCMatrix<T> right_matrix(vector);
+  CCMatrixArmadillo left_matrix(*this);
+  CCMatrixArmadillo right_matrix(vector);
+  
   // Store the size of both vectors to create a solution matrix with
   // the corresponding sizes
   // (First dimension for the left vector)
@@ -154,21 +173,20 @@ namespace scicellxx
    {
     n_values_right_vector = 1;
    }
-  // Create a zero matrix where to store the result
-  CCMatrix<T> solution(n_values_left_vector, n_values_right_vector);
+  // Create a zero vector where to store the result
+  CCMatrixArmadillo solution(n_values_left_vector, n_values_right_vector);
   // Perform the multiplication (this method is in charge of verifying
   // whether the matrices fulfill the requirements for matrix
   // multiplication)
   multiply_matrices(left_matrix, right_matrix, solution);
-  // Return the solution matrix
+  // Return the solution vector
   return solution;
  }
  
  // ===================================================================
  // Performs dot product with the current vector
  // ===================================================================
- template<class T>
- T CCVector<T>::dot(const CCVector &right_vector)
+ T CCVectorArmadillo::dot(const CCVectorArmadillo &right_vector)
  {
   // Check that THIS and the right vector have memory allocated
   if (!this->Is_own_memory_allocated || !right_vector.is_own_memory_allocated())
@@ -225,30 +243,20 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
   
-  // Get the vector pointer of the right vector
-  T *right_vector_pt = right_vector.vector_pt();
-  
   // Store the dot product of the vectors
-  T dot_product = 0.0;
-  
-  // Compute the dot product
-  for (unsigned long i = 0; i < n_values_this_vector; i++)
-   {
-    dot_product+= Vector_pt[i] * right_vector_pt[i];
-   }
-  
+  const T dot_product = arma::dot(*Arma_vector_pt, *(right_vector.arma_vector_pt()));
+  // Return the dot product
   return dot_product;
   
  }
  
  // ===================================================================
- // Transforms the input vector to a vector class type (virtual such
- // that each derived class has to implement it)
+ // Transforms the input vector to an armadillo vector class type
+ // (virtual such that each derived class has to implement it)
  // ===================================================================
- template<class T>
- void CCVector<T>::set_vector(const T *vector_pt,
-                              const unsigned long n,
-                              bool is_column_vector)
+ void CCVectorArmadillo::set_vector(const T *vector_pt,
+                                       const unsigned long n,
+                                       bool is_column_vector)
  {
   // Clean any possible previously allocated memory
   clean_up();
@@ -256,14 +264,43 @@ namespace scicellxx
   // Set the number of values
   this->NValues = n;
   
-  // Allocate memory for the vector
-  Vector_pt = new T[n];
+  // Create an Armadillo's matrix (makes an own copy of the data,
+  // therefore 'matrix_pt' may be deleted safely)
+  if (is_column_vector)
+   {
+    Arma_vector_pt = new arma::Mat<Real>(vector_pt, n, 1);
+   }
+  else
+   {
+    Arma_vector_pt = new arma::Mat<Real>(vector_pt, 1, n);
+   }
   
   // Mark the vector as allocated its own memory
   this->Is_own_memory_allocated = true;
   
-  // Copy the vector (an element by element copy, uff!!)
-  std::memcpy(Vector_pt, vector_pt, n*sizeof(T));
+  // Set the transposed status
+  this->set_as_column_vector(is_column_vector);
+  
+ }
+ 
+ // ===================================================================
+ // Receives an armadillo type Mat
+ // ===================================================================
+ void CCVectorArmadillo::set_vector(arma::Mat<Real> *arma_vector_pt,
+                                       const unsigned long n,
+                                       bool is_column_vector)
+ {
+  // Clean any possible previously allocated memory
+  clean_up();
+  
+  // Set the number of values
+  this->NValues = n;
+  
+  // Call the copy constructor of Armadillo
+  Arma_vector_pt = new arma::Mat<Real>(*arma_vector_pt);
+  
+  // Mark the matrix as having its own memory
+  this->Is_own_memory_allocated = true;
   
   // Set the transposed status
   this->set_as_column_vector(is_column_vector);
@@ -273,8 +310,7 @@ namespace scicellxx
  // ===================================================================
  // Clean up for any dynamically stored data
  // ===================================================================
- template<class T>
- void CCVector<T>::clean_up()
+ void CCVectorArmadillo::clean_up()
  {
   // Check whether the Vector allocated its own memory
   if (this->Is_own_memory_allocated)
@@ -287,7 +323,7 @@ namespace scicellxx
   else // if empty
    {
     // Set the pointer of the vector to NULL
-    Vector_pt = 0;
+    Arma_vector_pt = 0;
    }
   
  }
@@ -295,16 +331,15 @@ namespace scicellxx
  // ===================================================================
  // Free allocated memory for vector
  // ===================================================================
- template<class T>
- void CCVector<T>::free_memory_for_vector()
+ void CCVectorArmadillo::free_memory_for_vector()
  {
   // Is the vector allowed for deletion. If this method is called from
   // an external source we need to check whether the vector has been
   // marked for deletion
   if (this->Delete_vector)
    {
-    delete [] Vector_pt;
-    Vector_pt = 0; 
+    delete Arma_vector_pt;
+    Arma_vector_pt = 0; 
     
     // Mark the vector as not having memory allocated
     this->Is_own_memory_allocated=false;
@@ -326,9 +361,8 @@ namespace scicellxx
  // ===================================================================
  // Performs sum of vectors
  // ===================================================================
- template<class T>
- void CCVector<T>::add_vector(const CCVector<T> &vector,
-                              CCVector<T> &solution_vector)
+ void CCVectorArmadillo::add_vector(const CCVectorArmadillo &vector,
+                                       CCVectorArmadillo &solution_vector)
  {
   // Check that THIS and the other vector have memory allocated
   if (!this->Is_own_memory_allocated || !vector.is_own_memory_allocated())
@@ -401,24 +435,21 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
   // Get the vector pointer of the input vector
-  T *vector_pt = vector.vector_pt();
+  arma::Mat<Real> *arma_vector_pt = vector.arma_vector_pt();
+  
   // Perform the addition
-  for (unsigned long i = 0; i < n_values_this_vector; i++)
-   {
-    solution_vector_pt[i] = Vector_pt[i] + vector_pt[i];
-   }
+  (*arma_solution_vector_pt) = (*Arma_vector_pt) + (*arma_vector_pt);
   
  }
  
  // ===================================================================
  // Performs substraction of vectors
  // ===================================================================
- template<class T>
- void CCVector<T>::substract_vector(const CCVector<T> &vector,
-                                    CCVector<T> &solution_vector)
+ void CCVectorArmadillo::substract_vector(const CCVectorArmadillo &vector,
+                                    CCVectorArmadillo &solution_vector)
  {
   // Check that THIS and the other vector have no memory allocated
   if (!this->Is_own_memory_allocated || !vector.is_own_memory_allocated())
@@ -450,7 +481,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check that the three vectors have the same column vector status
   if (this->is_column_vector() != vector.is_column_vector() ||
       solution_vector.is_column_vector() != vector.is_column_vector())
@@ -491,25 +522,22 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
   // Get the vector pointer of the input vector
-  T *vector_pt = vector.vector_pt();
-  // Perform the addition
-  for (unsigned long i = 0; i < n_values_this_vector; i++)
-   {
-    solution_vector_pt[i] = Vector_pt[i] - vector_pt[i];
-   }
+  arma::Mat<Real> *arma_vector_pt = vector.arma_vector_pt();
   
+  // Perform the substraction
+  (*arma_solution_vector_pt) = (*Arma_vector_pt) - (*arma_vector_pt);
+    
  }
  
  // ===================================================================
  // Performs multiplication of vectors (element by element)
  // ===================================================================
- template<class T>
- void CCVector<T>::
- multiply_element_by_element_vector(const CCVector<T> &vector,
-                                    CCVector<T> &solution_vector)
+ void CCVectorArmadillo::
+ multiply_element_by_element_vector(const CCVectorArmadillo &vector,
+                                    CCVectorArmadillo &solution_vector)
  {
   // Check that THIS and the other vector have memory allocated
   if (!this->Is_own_memory_allocated || !vector.is_own_memory_allocated())
@@ -541,7 +569,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-  
+
   // Check that the three vectors have the same column vector status
   if (this->is_column_vector() != vector.is_column_vector() ||
       solution_vector.is_column_vector() != vector.is_column_vector())
@@ -582,23 +610,23 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
   // Get the vector pointer of the input vector
-  T *vector_pt = vector.vector_pt();
-  // Perform the addition
-  for (unsigned long i = 0; i < n_values_this_vector; i++)
+  arma::Mat<Real> *arma_vector_pt = vector.arma_vector_pt();
+  
+  // Perform the operation
+  for (unsigned long i = 0; i < this->NValues; i++)
    {
-    solution_vector_pt[i] = Vector_pt[i] * vector_pt[i];
+    (*arma_solution_vector_pt)(i) = (*Arma_vector_pt)(i) * (*arma_vector_pt)(i);
    }
   
  }
  
  // ===================================================================
- // Computes the transpose and store in the transposed vector
+ // Computes the transpose and store in the solution vector
  // ===================================================================
- template<class T>
- void CCVector<T>::transpose(CCVector<T> &transposed_vector)
+ void CCVectorArmadillo::transpose(CCVectorArmadillo &transposed_vector)
  {
   // Check that THIS vector has memory allocated
   if (!this->Is_own_memory_allocated)
@@ -613,18 +641,27 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
   
-  // Copy the vector into the tranposed vector
+  // Copy all the vector
   transposed_vector = (*this);
-  // Get the current "column vector" status of the vector and set the
-  // transposed status of the new vector
-  transposed_vector.set_as_column_vector(!(this->Is_column_vector));
+  // Perform tranposition
+  transposed_vector.transpose();  
+ }
+ 
+ // ===================================================================
+ // Transpose the vector
+ // ===================================================================
+ void CCVectorArmadillo::transpose()
+ {
+  // Performs the operation
+  arma::inplace_trans(*Arma_vector_pt);
+  // Change the status
+  this->Is_column_vector=!(this->Is_column_vector);
  }
  
  // ===================================================================
  // Get the specified value from the vector (read-only)
  // ===================================================================
- template<class T>
- const T CCVector<T>::value(const unsigned long i) const
+ const T CCVectorArmadillo::value(const unsigned long i) const
  {
 #ifdef SCICELLXX_RANGE_CHECK
   if (!(this->is_own_memory_allocated()))
@@ -651,14 +688,13 @@ namespace scicellxx
    } 
 #endif // #ifdef SCICELLXX_RANGE_CHECK
   // Return the value at position i
-  return Vector_pt[i];
+  return (*Arma_vector_pt)(i);
  }
  
  // ===================================================================
  // Set values in the vector (write version)
  // ===================================================================
- template<class T>
- T &CCVector<T>::value(const unsigned long i)
+ T &CCVectorArmadillo::value(const unsigned long i)
  {
 #ifdef SCICELLXX_RANGE_CHECK
   if (!(this->is_own_memory_allocated()))
@@ -685,14 +721,13 @@ namespace scicellxx
    } 
 #endif // #ifdef SCICELLXX_RANGE_CHECK
   // Return the value at row i and column j
-  return Vector_pt[i];
+  return (*Arma_vector_pt)(i);
  }
 
  // ===================================================================
  // Output the vector
  // ===================================================================
- template<class T>
- void CCVector<T>::output(bool output_indexes) const
+ void CCVectorArmadillo::output(bool output_indexes) const
  {
   if (!this->Is_own_memory_allocated)
    {
@@ -716,11 +751,7 @@ namespace scicellxx
      } // if (output_indexes)
     else
      {
-      for (unsigned long i = 0; i < this->NValues; i++)
-       {
-        std::cout << value(i) << " ";
-       } // for (i < this->NValues)
-      std::cout << std::endl;
+      Arma_vector_pt->print();
      } // else if (output_indexes)
    }
   
@@ -729,8 +760,7 @@ namespace scicellxx
  // ===================================================================
  // Output the vector
  // ===================================================================
- template<class T>
- void CCVector<T>::output(std::ofstream &outfile, bool output_indexes) const
+ void CCVectorArmadillo::output(std::ofstream &outfile, bool output_indexes) const
  {
   if (!this->Is_own_memory_allocated)
    {
@@ -754,32 +784,24 @@ namespace scicellxx
      } // if (output_indexes)
     else
      {
-      for (unsigned long i = 0; i < this->NValues; i++)
-       {
-        outfile << value(i) << " ";
-       } // for (i < this->NValues)
-      outfile << std::endl;
+      Arma_vector_pt->print(outfile);
      } // else if (output_indexes)
    }
   
  }
-
+ 
  // ===================================================================
  // Computes the norm-1 of the vector
  // ===================================================================
- template<class T>
- T CCVector<T>::norm_1()
+ T CCVectorArmadillo::norm_1()
  {
   // Sum
-  T sum = 0.0;
+  Real sum = 0.0;
   // Check whether the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
     // Compute the norm
-    for (unsigned long i = 0; i < this->NValues; i++)
-     {
-      sum+= std::fabs(Vector_pt[i]);
-     }
+    sum = arma::norm((*Arma_vector_pt), 1);
    }
   else
    {
@@ -800,19 +822,15 @@ namespace scicellxx
  // ===================================================================
  // Computes the norm-2 of the vector
  // ===================================================================
- template<class T>
- T CCVector<T>::norm_2()
+ T CCVectorArmadillo::norm_2()
  {
   // Sum
-  T sum = 0.0;
+  Real sum = 0.0;
   // Check whether the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
-    // Compute the norm 2
-    for (unsigned long i = 0; i < this->NValues; i++)
-     {
-      sum+= (Vector_pt[i]*Vector_pt[i]);
-     }
+    // Compute the norm
+    sum = arma::norm((*Arma_vector_pt), 2);
    }
   else
    {
@@ -826,29 +844,22 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
   
-  return sqrt(sum);
+  return sum;
   
  }
 
  // ===================================================================
  // Computes the infinite norm
  // ===================================================================
- template<class T>
- T CCVector<T>::norm_inf()
+ T CCVectorArmadillo::norm_inf()
  {
-  // Maximum
-  T norm = 0.0;
+  // Infinite norm
+  Real norm;
   // Check whether the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
-    // Initialise the maximum with the first value
-    norm = std::fabs(Vector_pt[0]);
     // Compute the norm
-    for (unsigned long i = 1; i < this->NValues; i++)
-     {
-      if (std::fabs(Vector_pt[i]) > norm)
-       norm = std::fabs(Vector_pt[i]);
-     }
+    norm = arma::norm((*Arma_vector_pt), "inf");
    }
   else
    {
@@ -863,28 +874,20 @@ namespace scicellxx
    }
   
   return norm;
-  
  }
  
  // ===================================================================
  // Computes the maximum value
  // ===================================================================
- template<class T>
- T CCVector<T>::max()
+ T CCVectorArmadillo::max()
  {
   // Maximum
-  T max = 0.0;
+  Real max;
   // Check whether the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
-    // Initialise the maximum with the first value
-    max = Vector_pt[0];
     // Compute the maximum
-    for (unsigned long i = 1; i < this->NValues; i++)
-     {
-      if (Vector_pt[i] > max)
-       max = Vector_pt[i];
-     }
+    max = Arma_vector_pt->max();
    }
   else
    {
@@ -899,28 +902,20 @@ namespace scicellxx
    }
   
   return max;
-  
  }
- 
+
  // ===================================================================
  // Computes the minimum value
  // ===================================================================
- template<class T>
- T CCVector<T>::min()
+ T CCVectorArmadillo::min()
  {
   // Minimum
-  T min = 0.0;
+  Real min;
   // Check whether the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
-    // Initialise the minimum with the first value
-    min = Vector_pt[0];
     // Compute the minimum
-    for (unsigned long i = 1; i < this->NValues; i++)
-     {
-      if (Vector_pt[i] < min)
-       min = Vector_pt[i];
-     }
+    min = Arma_vector_pt->min();
    }
   else
    {
@@ -935,25 +930,31 @@ namespace scicellxx
    }
   
   return min;
-  
  }
- 
+
  // ===================================================================
  // Allows to create a vector with the given size but with no data
  // ===================================================================
- template<class T>
- void CCVector<T>::allocate_memory(const unsigned long n)
+ void CCVectorArmadillo::allocate_memory(const unsigned long n)
  {
   // Clean any possibly stored data
   clean_up();
   
   // Set the number of rows and columns of the matrix
   this->NValues = n;
-
+  
+  // Allocate memory
   //allocate_memory();
   
   // Allocate memory for the vector
-  Vector_pt = new T[this->NValues];
+  if (this->Is_column_vector)
+   {    
+    Arma_vector_pt = new arma::Mat<Real>(this->NValues, 1);
+   }
+  else
+   {
+    Arma_vector_pt = new arma::Mat<Real>(1, this->NValues);
+   }
   
   // Mark the vector as allocated its own memory
   this->Is_own_memory_allocated=true;
@@ -963,30 +964,35 @@ namespace scicellxx
  // // ===================================================================
  // // Allocates memory to store entries of the vector
  // // ===================================================================
- // template<class T>
- // void CCVector<T>::allocate_memory()
+ // void CCVectorArmadillo::allocate_memory()
  // {
  //  // Delete any data in memory
  //  clean_up();
-  
+
  //  // Allocate memory for the vector
- //  Vector_pt = new T[this->NValues];
+ //  if (this->Is_column_vector)
+ //   {    
+ //    Arma_vector_pt = new arma::Mat<Real>(this->NValues, 1);
+ //   }
+ //  else
+ //   {
+ //    Arma_vector_pt = new arma::Mat<Real>(1, this->NValues);
+ //   }
   
  //  // Mark the vector as allocated its own memory
  //  this->Is_own_memory_allocated=true;
  // }
- 
+
  // ===================================================================
  // Fills the vector with zeroes
  // ===================================================================
- template<class T>
- void CCVector<T>::fill_with_zeroes()
+ void CCVectorArmadillo::fill_with_zeroes()
  {
   // Check that the vector has memory allocated
   if (this->Is_own_memory_allocated)
    {
     // Fill the vector with zeroes
-    std::memset(Vector_pt, 0, this->NValues*sizeof(T));
+    Arma_vector_pt->zeros();
    }
   else
    {
@@ -1011,8 +1017,7 @@ namespace scicellxx
  // ================================================================
  // Dot product of vectors
  // ================================================================
- template<class T>
- T dot_vectors(const CCVector<T> &left_vector, const CCVector<T> &right_vector)
+ T dot_vectors(const CCVectorArmadillo &left_vector, const CCVectorArmadillo &right_vector)
  {
   // Check that the left and the right vectors have memory allocated
   if (!left_vector.is_own_memory_allocated() || !right_vector.is_own_memory_allocated())
@@ -1069,20 +1074,9 @@ namespace scicellxx
                            SCICELLXX_EXCEPTION_LOCATION);
    }
   
-  // Get the vector pointer of the left vector
-  T *left_vector_pt = left_vector.vector_pt();
-  // Get the vector pointer of the right vector
-  T *right_vector_pt = right_vector.vector_pt();
-  
   // Store the dot product of the vectors
-  T dot_product = 0.0;
-  
-  // Compute the dot product
-  for (unsigned long i = 0; i < n_values_left_vector; i++)
-   {
-    dot_product+= left_vector_pt[i] * right_vector_pt[i];
-   }
-  
+  const T dot_product = arma::dot(*(left_vector.arma_vector_pt()), *(right_vector.arma_vector_pt()));
+  // Return the dot product
   return dot_product;
   
  }
@@ -1090,10 +1084,9 @@ namespace scicellxx
  // ================================================================
  // Addition of vectors
  // ================================================================
- template<class T>
- void add_vectors(const CCVector<T> &vector_one,
-                  const CCVector<T> &vector_two,
-                  CCVector<T> &solution_vector)
+ void add_vectors(const CCVectorArmadillo &vector_one,
+                  const CCVectorArmadillo &vector_two,
+                  CCVectorArmadillo &solution_vector)
  {
   // Check that the vectors have memory allocated
   if (!vector_one.is_own_memory_allocated() || !vector_two.is_own_memory_allocated())
@@ -1166,28 +1159,23 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
-  // Get the vector pointer of the vector one
-  T *vector_one_pt = vector_one.vector_pt();
-  // Get the vector pointer of the vector two
-  T *vector_two_pt = vector_two.vector_pt();
+  // Get the vector pointer of the input vectors
+  arma::Mat<Real> *arma_vector_one_pt = vector_one.arma_vector_pt();
+  arma::Mat<Real> *arma_vector_two_pt = vector_two.arma_vector_pt();
   
   // Perform the addition
-  for (unsigned long i = 0; i < n_values_vector_one; i++)
-   {
-    solution_vector_pt[i] = vector_one_pt[i] + vector_two_pt[i];
-   }
+  (*arma_solution_vector_pt) = (*arma_vector_one_pt) + (*arma_vector_two_pt);
   
  }
 
  // ================================================================
  // Substraction of vectors
  // ================================================================
- template<class T>
- void substract_vectors(const CCVector<T> &vector_one,
-                        const CCVector<T> &vector_two,
-                        CCVector<T> &solution_vector)
+ void substract_vectors(const CCVectorArmadillo &vector_one,
+                        const CCVectorArmadillo &vector_two,
+                        CCVectorArmadillo &solution_vector)
  {
   // Check that the vectors have no memory allocated
   if (!vector_one.is_own_memory_allocated() || !vector_two.is_own_memory_allocated())
@@ -1219,7 +1207,7 @@ namespace scicellxx
                            SCICELLXX_CURRENT_FUNCTION,
                            SCICELLXX_EXCEPTION_LOCATION);
    }
-
+  
   // Check that the three vectors have the same column vector status
   if (vector_one.is_column_vector() != vector_two.is_column_vector() ||
       solution_vector.is_column_vector() != vector_two.is_column_vector())
@@ -1260,28 +1248,23 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
-    
-  // Get the vector pointer of the vector one
-  T *vector_one_pt = vector_one.vector_pt();
-  // Get the vector pointer of the vector two
-  T *vector_two_pt = vector_two.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
-  // Perform the substraction of vectors
-  for (unsigned long i = 0; i < n_values_vector_one; i++)
-   {
-    solution_vector_pt[i] = vector_one_pt[i] - vector_two_pt[i];
-   }
+  // Get the vector pointer of the input vectors
+  arma::Mat<Real> *arma_vector_one_pt = vector_one.arma_vector_pt();
+  arma::Mat<Real> *arma_vector_two_pt = vector_two.arma_vector_pt();
+  
+  // Perform the addition
+  (*arma_solution_vector_pt) = (*arma_vector_one_pt) + (*arma_vector_two_pt);
   
  }
 
  // ================================================================
  // Performs multiplication of vectors (one by one entries)
  // ================================================================
- template<class T>
- void multiply_element_by_element_vectors(const CCVector<T> &vector_one,
-                                          const CCVector<T> &vector_two,
-                                          CCVector<T> &solution_vector)
+ void multiply_element_by_element_vectors(const CCVectorArmadillo &vector_one,
+                                          const CCVectorArmadillo &vector_two,
+                                          CCVectorArmadillo &solution_vector)
  {
   // Check that the vectors have memory allocated
   if (!vector_one.is_own_memory_allocated() || !vector_two.is_own_memory_allocated())
@@ -1289,9 +1272,9 @@ namespace scicellxx
     // Error message
     std::ostringstream error_message;
     error_message << "One of the vectors to operate with has no memory allocated\n"
-                  << "vector_one.is_own_memory_allocated() = "
+                  << "vector_one.is_own_memory_allocated(): "
                   << vector_one.is_own_memory_allocated() << "\n"
-                  << "vector_two.is_own_memory_allocated() = "
+                  << "vector_two.is_own_memory_allocated(): "
                   << vector_two.is_own_memory_allocated() << std::endl;
     throw SciCellxxLibError(error_message.str(),
                            SCICELLXX_CURRENT_FUNCTION,
@@ -1354,19 +1337,220 @@ namespace scicellxx
    }
   
   // Get the vector pointer of the solution vector
-  T *solution_vector_pt = solution_vector.vector_pt();
+  arma::Mat<Real> *arma_solution_vector_pt = solution_vector.arma_vector_pt();
   
-  // Get the vector pointer of the vector one
-  T *vector_one_pt = vector_one.vector_pt();
-  // Get the vector pointer of the vector two
-  T *vector_two_pt = vector_two.vector_pt();
-  
-  // Perform the substraction of vectors
+  // Get the vector pointer of the input vectors
+  arma::Mat<Real> *arma_vector_one_pt = vector_one.arma_vector_pt();
+  arma::Mat<Real> *arma_vector_two_pt = vector_two.arma_vector_pt();
+
+  // Perform the operation
   for (unsigned long i = 0; i < n_values_vector_one; i++)
    {
-    solution_vector_pt[i] = vector_one_pt[i] * vector_two_pt[i];
+    (*arma_solution_vector_pt)(i) = (*arma_vector_one_pt)(i) * (*arma_vector_two_pt)(i);
    }
   
  }
  
+ // ================================================================
+ // Concatenate vector horizontally
+ // ================================================================
+  void concatenate_vectors_horizontally(const CCVectorArmadillo &left_vector,
+                                        const CCVectorArmadillo &right_vector,
+                                        CCVectorArmadillo &concatenated_vector)
+ {
+  // Check that both vector have memory allocated
+  if (!left_vector.is_own_memory_allocated() || !right_vector.is_own_memory_allocated())
+   {
+    // Error message
+    std::ostringstream error_message;
+    error_message << "One of the vectors to operate with has no memory allocated\n"
+                  << "left_vector.is_own_memory_allocated(): "
+                  << left_vector.is_own_memory_allocated() << "\n"
+                  << "right_vector.is_own_memory_allocated(): "
+                  << right_vector.is_own_memory_allocated() << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
+   }
+  
+  // Check whether both vectors are row vectors
+  if (left_vector.is_column_vector() || right_vector.is_column_vector())
+   {
+    // Error message
+    std::ostringstream error_message;
+    error_message << "One of the vectors is not a row vector, if you need\n"
+                  << "to concatenate two row vectors to create a matrix\n"
+                  << "then BETTER FIRST transform your vectors to matrices\n"
+                  << "then concatenate them\n"
+                  << "left_vector.is_column_vector(): " << left_vector.is_column_vector()
+                  << "\nright_vector.is_column_vector(): " << right_vector.is_column_vector()
+                  << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
+   }
+  
+  // Check whether the concatenated vector is a row vector, if not
+  // then transpose it
+  if (concatenated_vector.is_column_vector())
+   {
+    // Warning message
+    std::ostringstream warning_message;
+    warning_message << "The resulting concatenated vector is a column vector\n"
+                    << "concatenated_vector.is_column_vector(): "
+                    << concatenated_vector.is_column_vector()
+                    << "\n\n Thus I am going to transpose it!!!"
+                    << std::endl;
+    SciCellxxLibWarning(warning_message.str(),
+                       SCICELLXX_CURRENT_FUNCTION,
+                       SCICELLXX_EXCEPTION_LOCATION);
+    
+    concatenated_vector.transpose();
+    
+   }
+  
+  // Get the number of elements of each vector and compute the new
+  // number of elements for the new vector
+  const unsigned long n_elements_left_vector = left_vector.n_values();
+  const unsigned long n_elements_right_vector = right_vector.n_values();
+  const unsigned long n_elements_new_concatenated_vector =
+   n_elements_left_vector + n_elements_right_vector;
+  
+  // Check whether the concatenated vector has allocated memory,
+  // otherwise allocate it here!!!
+  if (!concatenated_vector.is_own_memory_allocated())
+   {
+    // Allocate memory for the matrix (create a row vector)
+    concatenated_vector.allocate_memory(n_elements_new_concatenated_vector);
+   }
+  else
+   {
+    // Check that the already allocated memory is correct (n_values)
+    if (concatenated_vector.n_values() != n_elements_new_concatenated_vector)
+     {
+      // Error message
+      std::ostringstream error_message;
+      error_message << "The number of elements for the concatenated vector is\n"
+                    << "not as expected\n"
+                    << "dim(concatenated_vector): " << concatenated_vector.n_values()
+                    << "\nn_elements_new_concatenated_vector: " << n_elements_new_concatenated_vector
+                    << std::endl;
+      throw SciCellxxLibError(error_message.str(),
+                             SCICELLXX_CURRENT_FUNCTION,
+                             SCICELLXX_EXCEPTION_LOCATION);
+     }
+    
+   }
+  
+  // Get the matrix pointer of the concatenated matrix
+  arma::Mat<Real> *arma_concatenated_vector_pt = concatenated_vector.arma_vector_pt();
+  
+  // Get the vector pointer of the input vector
+  arma::Mat<Real> *arma_left_vector_pt = left_vector.arma_vector_pt();
+  arma::Mat<Real> *arma_right_vector_pt = right_vector.arma_vector_pt();
+  
+  // Perform the concatenation of rows/horizontal_concatenation
+  (*arma_concatenated_vector_pt) = arma::join_rows((*arma_left_vector_pt), (*arma_right_vector_pt));
+  
+ }
+ 
+ // ================================================================
+ // Concatenate matrices vertically
+ // ================================================================
+  void concatenate_vectors_vertically(const CCVectorArmadillo &upper_vector,
+                                      const CCVectorArmadillo &lower_vector,
+                                      CCVectorArmadillo &concatenated_vector)
+ {
+  // Check that both vector have memory allocated
+  if (!upper_vector.is_own_memory_allocated() || !lower_vector.is_own_memory_allocated())
+   {
+    // Error message
+    std::ostringstream error_message;
+    error_message << "One of the vectors to operate with has no memory allocated\n"
+                  << "upper_vector.is_own_memory_allocated(): "
+                  << upper_vector.is_own_memory_allocated() << "\n"
+                  << "lower_vector.is_own_memory_allocated(): "
+                  << lower_vector.is_own_memory_allocated() << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
+   }
+  
+  // Check whether both vectors are column vectors
+  if (!upper_vector.is_column_vector() || !lower_vector.is_column_vector())
+   {
+    // Error message
+    std::ostringstream error_message;
+    error_message << "One of the vectors is not a column vector, if you need\n"
+                  << "to concatenate two column vectors to create a matrix\n"
+                  << "then BETTER FIRST transform your vectors to matrices\n"
+                  << "then concatenate them\n"
+                  << "upper_vector.is_column_vector(): " << upper_vector.is_column_vector()
+                  << "\nlower_vector.is_column_vector(): " << lower_vector.is_column_vector()
+                  << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
+   }
+  
+  // Check whether the concatenated vector is a column vector
+  if (!concatenated_vector.is_column_vector())
+   {
+    // Error message
+    std::ostringstream error_message;
+    error_message << "The resulting concatenated vector is a row vector\n"
+                  << "concatenated_vector.is_column_vector(): "
+                  << concatenated_vector.is_column_vector()
+                  << std::endl;
+    throw SciCellxxLibError(error_message.str(),
+                           SCICELLXX_CURRENT_FUNCTION,
+                           SCICELLXX_EXCEPTION_LOCATION);
+   }
+  
+  // Get the number of elements of each vector and compute the new
+  // number of elements for the new vector
+  const unsigned long n_elements_upper_vector = upper_vector.n_values();
+  const unsigned long n_elements_lower_vector = lower_vector.n_values();
+  const unsigned long n_elements_new_concatenated_vector =
+   n_elements_upper_vector + n_elements_lower_vector;
+  
+  // Check whether the concatenated vector has allocated memory,
+  // otherwise allocate it here!!!
+  if (!concatenated_vector.is_own_memory_allocated())
+   {
+    // Allocate memory for the matrix (create a column vector)
+    concatenated_vector.allocate_memory(n_elements_new_concatenated_vector);
+   }
+  else
+   {
+    // Check that the already allocated memory is correct (n_values)
+    if (concatenated_vector.n_values() != n_elements_new_concatenated_vector)
+     {
+      // Error message
+      std::ostringstream error_message;
+      error_message << "The number of elements for the concatenated vector is\n"
+                    << "not as expected\n"
+                    << "dim(concatenated_vector): " << concatenated_vector.n_values()
+                    << "\nn_elements_new_concatenated_vector: " << n_elements_new_concatenated_vector
+                    << std::endl;
+      throw SciCellxxLibError(error_message.str(),
+                             SCICELLXX_CURRENT_FUNCTION,
+                             SCICELLXX_EXCEPTION_LOCATION);
+     }
+    
+   }
+  
+  // Get the matrix pointer of the concatenated matrix
+  arma::Mat<Real> *arma_concatenated_vector_pt = concatenated_vector.arma_vector_pt();
+  
+  // Get the vector pointer of the input vector
+  arma::Mat<Real> *arma_upper_vector_pt = upper_vector.arma_vector_pt();
+  arma::Mat<Real> *arma_lower_vector_pt = lower_vector.arma_vector_pt();
+  
+  // Perform the concatenation of rows/horizontal_concatenation
+  (*arma_concatenated_vector_pt) = arma::join_cols((*arma_upper_vector_pt), (*arma_lower_vector_pt));
+  
+ }
+ 
 }
+
